+++ .github/workflows/integration-tests.yml
@@ -120,7 +120,6 @@ jobs:
       fail-fast: false
       matrix:
         test_type: [
-          "Cork",
           "ScheduledCork",
           "AuctionModule",
           "CellarFees",
